cinema.sh - updated to work with the refactored website
This script fetches its data from Cinecartaz, but the website changed last
august. This commit now updates the parser to scap the data as it shows on the
webpage now. Unicode entities are being used on the movies' titles (instead of
HTML entities) which made the script a little bit messier.